apps/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
anthropic==0.7.0
beautifulsoup4==4.12.2
requests==2.31.0
pydantic[email]==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
supabase==2.3.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
PyJWT==2.8.0
email-validator==2.1.0